# internal/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CourseProgressResponse:
    properties:
      completed_lesson_ids:
        items:
          type: string
        type: array
      completed_lessons:
        type: integer
      course_id:
        type: string
      progress_percent:
        type: number
      total_lessons:
        type: integer
    type: object
...
  dto.CourseStatisticsResponse:
    properties:
      course_id:
        type: string
      users:
        items:
          $ref: '#/definitions/dto.UserCourseProgressResponse'
        type: array
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.LessonProgressListResponse:
    properties:
      progress:
        items:
          $ref: '#/definitions/dto.LessonProgressResponse'
        type: array
    type: object
  dto.LessonProgressRequest:
    properties:
      lesson_id:
        type: string
    type: object
  dto.LessonProgressResponse:
    properties:
      lesson_id:
        type: string
      user_id:
        type: string
    type: object
...
  dto.ModuleListResponse:
    properties:
      modules:
        items:
          $ref: '#/definitions/dto.ModuleResponse'
        type: array
    type: object
  dto.ModuleProgressResponse:
    properties:
      completed_lesson_ids:
        items:
          type: string
        type: array
      completed_lessons:
        type: integer
      module_id:
        type: string
      progress_percent:
        type: number
      total_lessons:
        type: integer
    type: object
  dto.ModuleRequest:
    properties:
      course_id:
        type: string
      description:
        type: string
      position:
        type: integer
      title:
        type: string
    type: object
  dto.ModuleResponse:
    properties:
      amount_of_lessons:
        type: integer
      course_id:
        type: string
      description:
        type: string
      id:
        type: string
      position:
        type: integer
      title:
        type: string
    type: object
  dto.ModuleStatisticsResponse:
    properties:
      module_id:
        type: string
      users:
        items:
          $ref: '#/definitions/dto.UserModuleProgressResponse'
        type: array
    type: object
...
  dto.UserCourseProgressResponse:
    properties:
      completed:
        type: boolean
      completed_lessons:
        type: integer
      progress_percent:
        type: number
      total_lessons:
        type: integer
      user_id:
        type: string
    type: object
...
  dto.UserModuleProgressResponse:
    properties:
      completed:
        type: boolean
      completed_lessons:
        type: integer
      progress_percent:
        type: number
      total_lessons:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API для пользователей Course Service
  title: Course Service API
  version: "1.0"
paths:
...
  /lesson-progresses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LessonProgressListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить прогресс пользователя по урокам
      tags:
      - Lesson Progress
    post:
      consumes:
      - application/json
      parameters:
      - description: Lesson ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LessonProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LessonProgressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отметить урок как пройденный
      tags:
      - Lesson Progress
  /lesson-progresses/lessons/{lessonId}:
    get:
      parameters:
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LessonProgressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить прогресс по уроку
      tags:
      - Lesson Progress
...
  /modules:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные модуля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ModuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать модуль
      tags:
      - Modules
  /modules/{id}:
    delete:
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить модуль
      tags:
      - Modules
    get:
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить модуль
      tags:
      - Modules
    patch:
      consumes:
      - application/json
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: string
      - description: Новые данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ModuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить модуль
      tags:
      - Modules
  /modules/courses/{course_id}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModuleListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить модули курса
      tags:
      - Modules
  /progresses/courses/{courseId}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CourseProgressResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить прогресс курса
      tags:
      - Course Progress
  /progresses/courses/{courseId}/statistics:
    get:
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CourseStatisticsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить статистику курса
      tags:
      - Course Progress
  /progresses/modules/{moduleId}:
    get:
      parameters:
      - description: Module ID
        in: path
        name: moduleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModuleProgressResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить прогресс модуля
      tags:
      - Module Progress
  /progresses/modules/{moduleId}/statistics:
    get:
      parameters:
      - description: Module ID
        in: path
        name: moduleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModuleStatisticsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить статистику модуля
      tags:
      - Module Progress
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"